Started a Jupyter Notebook
Jupyter notebook will be used as the main, shared project diary.
We decided to use Jupyter notebook as a diary as the project is largely computational, and somewhat mathematical, Jupyter notebook handling code and TeX is best suited. We will use a OneNote diary only to share link and pdf sources.