Corrections and more plots
I corrected the factor of 10^6. At least I'm quite sure that that is what caused it. The distance modulus equation already required d_L to be in Mpc, which I then converted to Mpc again when plotting it.
I've put the simulated models to their own file, keeping it clean will be important with this project.
I cleaned up the code a bit too, but didn't yet get the chance to try to vectorise the model calculation. Will give that a go soon.

I didn't change the mistake in the diary for the record, just inserted a warning and a new figure at today's dated work.
@ALonelylepton