corrected constant OL chi^2 analysis
It didn't work because you @ALonelylepton only appended to chi^2 array at the end and in one of the if statements. Also made it much neater, all these loops were unnecessary. On top of that, you forgot to include the changing of S_omega_k from linear to sin to sinh, and when we do that, we can also no longer cancel out the sqrt(Omega_k) terms, so I include these.
All described in the diary.