Small start on including H0 overall
Made a tiny start on including H0 as a parameter. Didn't have much time for this.
I looped over H0 as well as Om, but its very slow (takes about 5mins to run). And seems not to give expected results (all rows are the same, as if one of the parameters doesn't make a difference).