Much new, read push description
Added a dataset of 500 by 500 points for H(z) to match SNe accuracies. Then I combined both datasets to draw overall constraints. Everything is well documented in the Diary. The results are pretty stupid though. We still have the over constraining issue. We have to find it's root now. Especially that it crept into our different model. It has to be likelihoods now, surely. Also, it looks like our chi^2 for H(z) isn't overconstrained, it shows a decent degeneracy and has a large spread in H_0.