Fix to likelihoods constrained tighter than chi^2
Still in the process of updating the Diary with all this, but the codes should all be ready.
Will continue on this tonight, more pushed to come still today.
The H(z) data depending weakly on H_0 might not be an issue, it should be further discussed if time allows.
The SNe constraints in Om and H_0 are likely still too tight in H_0. That issue has not been fully resolved yet.
Will update official repository issues soon.