Opened New Branch
Opened a new branch for updates post submission deadline.
This branch is made so that any further work on this project can continue if needed after the submission deadline for the project as part of author's Bachelor's work. This way the marking can proceed fairly, while the project isn't required to end.
@ALonelylepton